+++ public/js/app.js
@@ -4,16 +4,16 @@
 //           notifications, volume sliders, status bar
 // ═══════════════════════════════════════════════════════════
 
-import SocketMethods   from './modules/app-socket.js?v=4.9.1';
-import UIBindMethods   from './modules/app-ui.js?v=4.10.0';
-import MediaMethods    from './modules/app-media.js?v=4.10.0';
+import SocketMethods   from './modules/app-socket.js?v=4.10.1';
+import UIBindMethods   from './modules/app-ui.js?v=4.10.1';
+import MediaMethods    from './modules/app-media.js?v=4.10.1';
 import ContextMethods  from './modules/app-context.js?v=4.9.1';
 import ChannelMethods  from './modules/app-channels.js?v=4.9.1';
-import MessageMethods  from './modules/app-messages.js?v=4.9.1';
+import MessageMethods  from './modules/app-messages.js?v=4.10.1';
 import UserMethods     from './modules/app-users.js?v=4.9.1';
 import VoiceMethods    from './modules/app-voice.js?v=4.7.1';
 import UtilityMethods  from './modules/app-utilities.js?v=4.10.0';
-import AdminMethods    from './modules/app-admin.js?v=4.10.0';
+import AdminMethods    from './modules/app-admin.js?v=4.10.1';
 import PlatformMethods from './modules/app-platform.js?v=4.9.2';
 import SearchMethods   from './modules/app-search.js?v=4.9.1';
 import FerryMethods    from './modules/app-ferry.js?v=3.51.4';