+++ public/js/modules/app-admin.js
@@ -3428,7 +3428,8 @@ _uploadGeneralFile(file, targetCode) {
         code,
         content,
         replyTo: (code === this.currentChannel && this.replyingTo) ? this.replyingTo.id : null,
-        ...(file && file._tags && file._tags.length ? { attachmentTags: file._tags } : {})
+        ...(file && file._tags && file._tags.length ? { attachmentTags: file._tags } : {}),
+        ...(file && file._selfDestruct ? { selfDestructMinutes: file._selfDestruct } : {})
       });
       if (file && file._tags && file._tags.length) this._recordFrequentTags(file._tags);
       this.notifications.play('sent');